requirements/requirements.txt
This are the requirements needed to run this project:

certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
panda==0.3.1
requests==2.31.0
setuptools==68.2.2
urllib3==2.0.7
